Refactor remote connector to make it easy to extends
